+++ CHANGES.rst
@@ -14,9 +14,9 @@ Version 2.0.26
 
 Released on July 19th 2017.
 
-- Description-extractor dependency updated to version 2.35 (`#
+- Description-extractor dependency updated to version 2.35 (`#429
   <https://github.com/mutalyzer/mutalyzer/pull/429>`_).
-- Fix for negative cCDSStop due to wrong transcript to protein link. (`#
+- Fix for negative cCDSStop due to wrong transcript to protein link. (`#430
   <https://github.com/mutalyzer/mutalyzer/issues/430>`_)